==> 2 DATA Cleaning.ipynb <==
import pandas as pd
import numpy as np
import re

samsung = pd.read_csv(r"C:\Users\YASH KHADTARE\FERGUSSON\SEM 3\Final year Project\1 DATA COLLECTION\SAMSUNG.csv")
samsung.head(5)

vivo = pd.read_csv(r"C:\Users\YASH KHADTARE\FERGUSSON\SEM 3\Final year Project\1 DATA COLLECTION\VIVO.csv")
realme = pd.read_csv(r"C:\Users\YASH KHADTARE\FERGUSSON\SEM 3\Final year Project\1 DATA COLLECTION\REALME.csv")
poco = pd.read_csv(r"C:\Users\YASH KHADTARE\FERGUSSON\SEM 3\Final year Project\1 DATA COLLECTION\POCO.csv")
oppo = pd.read_csv(r"C:\Users\YASH KHADTARE\FERGUSSON\SEM 3\Final year Project\1 DATA COLLECTION\OPPO.csv")
moto_google_nothing = pd.read_csv(r"C:\Users\YASH KHADTARE\FERGUSSON\SEM 3\Final year Project\1 DATA COLLECTION\Nothing,MOTO,Google.csv")
mi_redmi_xiaomi = pd.read_csv(r"C:\Users\YASH KHADTARE\FERGUSSON\SEM 3\Final year Project\1 DATA COLLECTION\MI , Redmi , Xiaomi.csv")
iphone = pd.read_csv(r"C:\Users\YASH KHADTARE\FERGUSSON\SEM 3\Final year Project\1 DATA COLLECTION\Iphone.csv")

df = pd.concat([samsung,vivo,realme,poco,oppo,moto_google_nothing,mi_redmi_xiaomi,iphone] , axis = 0)
df

# df = df.sample(10000)

df.drop("Unnamed: 0" , axis =1 , inplace=True)                                                                                       

df.reset_index(drop=True, inplace=True)
df.head(10)

df.info()

df.isnull().sum()

# 196 null values


df.dropna(inplace = True)
df.isnull().sum()

def remove_punctuations(i):
    
    i=re.sub(r'[^a-zA-Z ]','',i)
    
    return i

df['cleaned_review'] = df['review'].apply(remove_punctuations)

df.head()

# Seperating Joined words ,  ex - "Hinge is good and phone is handyCG glass Victus is goodDisplay is goodNo fast charging"
# These occur due to line break or paragraph change
def handle_joined_words(i):
    i = re.sub(r'([a-z])([A-Z])', r'\1 \2', i)
    return i


df['cleaned_review'] = df['cleaned_review'].apply(handle_joined_words)

print(df.iloc[1 , 1] ,"\n", df.iloc[1 , -1])

# all reviews to Lower case

df['cleaned_review'] = df['cleaned_review'].str.lower()

df

# Lemmatization

import nltk
from nltk.corpus import stopwords
from nltk.stem import WordNetLemmatizer
from nltk.corpus import wordnet

word_net = WordNetLemmatizer()        #creating word net lemmatizing object


def pos_tagger(nltk_tag):
    if nltk_tag.startswith('J'):
        return wordnet.ADJ
    elif nltk_tag.startswith('V'):
        return wordnet.VERB
    elif nltk_tag.startswith('N'):
        return wordnet.NOUN
    elif nltk_tag.startswith('R'):
        return wordnet.ADV
    else:         
        return None

def lemmatizing(i):
    words_split = i.split()   #splitting words
    
    no_stopwords = [i for i in words_split if i not in stopwords.words("English")]  #removing stopwords
     #Giving part of speech to each word
    no_stopwords = nltk.pos_tag(no_stopwords)      
    
    List=[]                   
    
    for i in no_stopwords:
        a=(i[0],pos_tagger(i[1]))
        List.append(a)


    # apply pos tag and lemmatization
    list_of_strings = []
    for i in List:
        try: 
            list_of_strings.append(word_net.lemmatize(i[0],i[1]))
        except:
            list_of_strings.append(word_net.lemmatize(i[0]))
    return " ".join(list_of_strings)


x = "Hinge is good and phone is handy CG glass Victus is good Display is good No fast charging"
lemmatizing(x)

import time
START = time.time()
df['cleaned_review'] = df['cleaned_review'].apply(lemmatizing)
STOP = time.time()
print((STOP - START)/60)

df

DATA = df
DATA.to_csv("DATA.csv")
