==> 5_Segmenting.ipynb <==
import pandas as pd
import numpy as np
import re

df = pd.read_csv(r"/content/drive/MyDrive/DATASETS/DATA.csv")
df.drop('Unnamed: 0',axis=1,inplace=True)
df.head(5)

def extract_company_name(text):
   return (re.match(r'^\w+',text).group(0)).upper()

df['Brand_name'] = df['model_name'].apply(extract_company_name)

**For transformers we dont need the cleaned review as it is lemmatized , so we drop that column**

df.drop('cleaned_review',axis=1,inplace=True)

# Rearranging columns
new_order = ['Brand_name','model_name','price','review','rating_by_reviewer']
df = df[new_order]
df.head(5)

**Now we divide the datasets into 3 parts , as we are 3 team members and each one run the created pipeline of sentimental analyis on their colab account , this will save our time (we will simultaneoulsy run all 3 notebooks)**

int(len(df)/3)

DATA_YASH = df.iloc[:118293,:]
DATA_ANURADHA = df.iloc[118293:(118293*2),:]
DATA_MAHESH = df.iloc[(118293*2):,:]

DATA_YASH.to_csv("DATA_YASH.csv")
DATA_ANURADHA.to_csv("DATA_ANURADHA.csv")
DATA_MAHESH.to_csv("DATA_MAHESH.csv")
