==> 6_Predictions_YASH_(Pipeline).ipynb <==
import time
START = time.time()

# **Now we will Create a final pipeline which will contain all the preprocessing and the model application**

# Libraries
# !pip install transformers
# !pip install dill
 

def remove_punctuations(texts):
  output = []
  for text in texts :
    x = re.sub(r'[^a-zA-Z ]','',text)
    output.append(x)
  return output

def handle_joined_words(texts):
  output = []
  for text in texts:
    x = re.sub(r'([a-z])([A-Z])', r'\1 \2', text)
    output.append(x)
  return output


# Creating function transformer

remove_punctuations = FunctionTransformer(remove_punctuations,validate = False)
handle_joined_words = FunctionTransformer(handle_joined_words,validate = False)


# Creating CardiffNLPClassifier OBJECT
class CardiffNLPClassifier:
    def __init__(self):
        self.device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
        self.tokenizer = AutoTokenizer.from_pretrained("cardiffnlp/twitter-roberta-base-sentiment-latest")
        self.config = AutoConfig.from_pretrained("cardiffnlp/twitter-roberta-base-sentiment-latest")
        self.model = AutoModelForSequenceClassification.from_pretrained("cardiffnlp/twitter-roberta-base-sentiment-latest").to(self.device)

    def transform(self, texts):
        results = []
        i = 0
        for text in texts:
            encoded_input = self.tokenizer(text, return_tensors='pt')
            encoded_input = {key: tensor.to(self.device) for key, tensor in encoded_input.items()}

            with torch.no_grad():
                output = self.model(**encoded_input)

            scores = output.logits.squeeze().detach().cpu().numpy()
            scores = softmax(scores)
            ranking = np.argsort(scores)[::-1]

            # Creating a dictionary with labels and probabilities
            d = {self.config.id2label[ranking[i]].upper(): np.round(float(scores[ranking[i]]), 4) for i in range(scores.shape[0])}

            # Removing 'neutral' class
            del d['NEUTRAL']

            # Extracting label with the highest probability
            key_with_highest_value = max(d, key=d.get)
            results.append(key_with_highest_value)
            if i % 100 == 0 :
              print(i)
            i = i + 1
        return results

sentiment_classifier = CardiffNLPClassifier()


# class CardiffNLPClassifier:
#     def __init__(self):
#         self.tokenizer = AutoTokenizer.from_pretrained("cardiffnlp/twitter-roberta-base-sentiment-latest")
#         self.config = AutoConfig.from_pretrained("cardiffnlp/twitter-roberta-base-sentiment-latest")
#         self.model = AutoModelForSequenceClassification.from_pretrained("cardiffnlp/twitter-roberta-base-sentiment-latest")

#     # def fit(self, texts, y=None):
#     #     return self

#     def transform(self, texts):
#         results = []
#         i = 0
#         for text in texts:
#             encoded_input = self.tokenizer(text, return_tensors='pt')
#             output = self.model(**encoded_input)
#             scores = output[0][0].detach().numpy()
#             scores = softmax(scores)
#             ranking = np.argsort(scores)[::-1]

#             # Creating a dictionary with labels and probabilities
#             d = {self.config.id2label[ranking[i]].upper(): np.round(float(scores[ranking[i]]), 4) for i in range(scores.shape[0])}

#             # Removing 'neutral' class
#             del d['NEUTRAL']

#             # Extracting label with the highest probability
#             key_with_highest_value = max(d, key=d.get)
#             results.append(key_with_highest_value)
#             if (i % 100) == 0:
#               print(i)
#             i = i+1
#         return results

# sentiment_classifier = CardiffNLPClassifier()

# Example
sentiment_analysis_pipeline = Pipeline([('remove_punctuations',remove_punctuations),('handle_joined_words',handle_joined_words),('sentiment_classifier',sentiment_classifier)])
X = ['lovely !!!','terrible','WOW I JUSTLOVED THIS','service is fine']
processed_text = sentiment_analysis_pipeline.transform(X)
print(processed_text)

# saving in form of pickle file
with open('/content/drive/MyDrive/DATASETS/sentiment_analysis_pipeline.pkl', 'wb') as f:
    dill.dump(sentiment_analysis_pipeline, f)

df = pd.read_csv(r"C:\Users\YASH KHADTARE\FERGUSSON\SEM 3\Final year Project\DATA_YASH.csv")
df = df.sample(500)
df.drop('Unnamed: 0',axis=1,inplace=True)
df

# Predictions
df['Sentiment'] = sentiment_analysis_pipeline.transform(df['review'])

df.head(5)

#Saving DATAFRAME
df.to_csv('/content/drive/MyDrive/DATASETS/FINAL_DATA_YASH.csv', index = False)

STOP = time.time()
print("Time required to run notebook : " , (STOP-START)/60, " mins")
